% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{American.Option.Value}
\alias{American.Option.Value}
\title{N-Factor Model American Put Option Pricing}
\usage{
American.Option.Value(
  X.0,
  parameters,
  n,
  t,
  dt,
  K,
  r,
  orthogonal = "Power",
  degree = 2,
  verbose = FALSE,
  debugging = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

American.Option.Value(X.0 = c(3,0),
                     parameters = SS_oil$two_factor,
                     n = 1e2,
                     t = 1,
                     dt = 1/12,
                     K = 20,
                     r = 0.05,
                     verbose = FALSE,
                     orthogonal = "Power",
                     degree = 2)

## ->

American_option_value(x_0 = c(3,0),
                     parameters = SS_oil$two_factor,
                     N_simulations = 1e2,
                     t = 1,
                     dt = 1/12,
                     K = 20,
                     r = 0.05,
                     verbose = FALSE,
                     orthogonal = "Power",
                     degree = 2)
}
\keyword{internal}
