% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{European.Option.Value}
\alias{European.Option.Value}
\title{N-Factor Model European Option Pricing}
\usage{
European.Option.Value(X.0, parameters, t, TTM, K, r, call, verbose = FALSE)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

##Step 2 - Calculate 'call' option price:
output <-  European.Option.Value(X.0 = c(3,0),
                      parameters = SS.Oil$Two.Factor,
                      t = 1,
                      TTM = 1,
                      K = 20,
                      r = 0.05,
                      call = TRUE,
                      verbose = FALSE)

## ->

output <- European_option_value(x_0 = c(3,0),
                     parameters = SS_oil$two_factor,
                     futures_maturity = 1,
                     option_maturity = 1,
                     K = 20,
                     r = 0.05,
                     call = TRUE,
                     verbose = FALSE)

}
\keyword{internal}
