% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Futures.Price.Forecast}
\alias{Futures.Price.Forecast}
\title{Forecast N-Factor Model Futures Prices}
\usage{
Futures.Price.Forecast(X.0, parameters, t = 0, TTM = 1:10, Percentiles = NULL)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

output <- Futures.Price.Forecast(X.0 = c(3,0),
                     parameters = SS.Oil$Two.Factor,
                     t = 0,
                     TTM = seq(0,9,1/12),
                     Percentiles = c(0.1, 0.9))

## ->

output <- futures_price_forecast(x_0 = c(3,0),
                      parameters = SS_oil$two_factor,
                      t = 0,
                      futures_TTM = seq(0,9,1/12),
                      percentiles = c(0.1, 0.9))

}
\keyword{internal}
