% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Futures.Price.Simulate}
\alias{Futures.Price.Simulate}
\title{Simulate N-Factor Model Futures Prices}
\usage{
Futures.Price.Simulate(X.0, parameters, dt, N.obs, TTM, verbose = TRUE)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

output <- Futures.Price.Simulate(X.0 = c(log(SS.Oil$Spot[1,1]), 0),
                     parameters = SS.Oil$Two.Factor,
                     dt = SS.Oil$dt,
                     N.obs = nrow(SS.Oil$Stitched.Futures),
                     TTM = SS.Oil$Stitched.TTM)

## ->

output <- futures_price_simulate(x_0 = c(log(SS_oil$spot[1,1]), 0),
                     parameters = SS_oil$two_factor,
                     dt = SS_oil$dt,
                     N_obs = nrow(SS_oil$stitched_futures),
                     futures_TTM = SS_oil$stitched_TTM)
}
\keyword{internal}
