% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{NFCP.Domains}
\alias{NFCP.Domains}
\title{N-Factor MLE Search Boundaries}
\usage{
NFCP.Domains(
  parameters,
  kappa = NULL,
  lambda = NULL,
  sigma = NULL,
  mu = NULL,
  mu_star = NULL,
  rho = NULL,
  contract = NULL,
  X.0 = NULL,
  E = NULL
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

output <- NFCP.Domains(names(SS_oil$two_factor))

## ->

output <- NFCP_domains(names(SS_oil$two_factor))

}
\keyword{internal}
