% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{NFCP.Kalman.filter}
\alias{NFCP.Kalman.filter}
\title{N-Factor Commodity Pricing Kalman Filter}
\usage{
NFCP.Kalman.filter(
  parameter.values,
  parameters,
  log.futures,
  dt,
  TTM,
  verbose = FALSE,
  debugging = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

output <- NFCP.Kalman.filter(
 parameter.values = SS.Oil$Two.Factor,
 parameters = names(SS.Oil$Two.Factor),
 log.futures = log(SS_oil$stitched_futures),
 TTM = SS_oil$stitched_TTM,
 dt = SS_oil$dt,
 verbose = FALSE)

## ->

output <- NFCP_Kalman_filter(
parameter_values = SS_oil$two_factor,
parameter_names = names(SS_oil$two_factor),
log_futures = log(SS_oil$stitched_futures),
futures_TTM = SS_oil$stitched_TTM,
dt = SS_oil$dt,
verbose = FALSE)

}
\keyword{internal}
