% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{NFCP.MLE}
\alias{NFCP.MLE}
\title{N-Factor Model Parameter Estimation through the Kalman Filter and Maximum Likelihood Estimation}
\usage{
NFCP.MLE(
  log.futures,
  dt,
  TTM,
  N.factors,
  GBM = TRUE,
  S.Constant = TRUE,
  Estimate.Initial.State = FALSE,
  Richardsons.Extrapolation = TRUE,
  cluster = FALSE,
  Domains = NULL,
  ...
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

SS_2F <- NFCP.MLE(
 ####Arguments
 log.futures = log(SS_oil$stitched_futures)[1:5,1],
 dt = SS_oil$dt,
 TTM= SS_oil$stitched_TTM[1],
 S.Constant = FALSE, N.factors = 1, GBM = TRUE,
 ####Genoud arguments:
 hessian = TRUE,
 Richardsons.Extrapolation = FALSE,
 pop.size = 4, optim.method = "L-BFGS-B", print.level = 0,
 max.generations = 0, solution.tolerance = 10)

## ->
NFCP_MLE(
####Arguments
log_futures = log(SS_oil$contracts)[1:20,1:5],
dt = SS_oil$dt,
futures_TTM= SS_oil$contract_maturities[1:20,1:5],
fixed_ME = TRUE,
N_factors = 1, GBM = TRUE,
####Genoud arguments:
hessian = TRUE,
Richardsons_extrapolation = FALSE,
pop.size = 4, optim.method = "L-BFGS-B", print.level = 0,
max.generations = 0, solution.tolerance = 10)


}
\keyword{internal}
