% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{NFCP.Parameters}
\alias{NFCP.Parameters}
\title{Specify parameters of N-factor model}
\usage{
NFCP.Parameters(
  N.factors,
  GBM,
  Initial.State,
  S.Constant,
  N.contracts = NULL,
  verbose = TRUE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{
output <- NFCP.Parameters(N.factors = 2,
               GBM = TRUE,
               Initial.State = FALSE,
               S.Constant = FALSE,
               N.contracts = 5)

## ->

output <- NFCP_parameters(N_factors = 2,
               GBM = TRUE,
               initial_states = FALSE,
               N_ME = 5)
}
\keyword{internal}
