% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Spot.Price.Forecast}
\alias{Spot.Price.Forecast}
\title{Forecast N-Factor Model Spot Prices}
\usage{
Spot.Price.Forecast(X.0, parameters, t, Percentiles = NULL)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

Spot.Price.Forecast(X.0 = c(3,0),
                  parameters = SS_oil$two_factor,
                  t = seq(0,9,1/12),
                  Percentiles = c(0.1, 0.9))

## ->

spot_price_forecast(x_0 = c(3,0),
                   parameters = SS_oil$two_factor,
                   t = seq(0,9,1/12),
                   percentiles = c(0.1, 0.9))

}
\keyword{internal}
