% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Spot.Price.Simulate}
\alias{Spot.Price.Simulate}
\title{Simulate N-Factor Model Spot Prices}
\usage{
Spot.Price.Simulate(
  X.0,
  parameters,
  t = 1,
  dt = 1,
  n = 2,
  antithetic = TRUE,
  verbose = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

output <- Spot.Price.Simulate(
X.0 = c(3,0),
parameters = SS.Oil$Two.Factor,
t = 1,
dt = 1/12,
n = 1e3,
antithetic = TRUE,
verbose = TRUE)

## ->

output <- spot_price_simulate(
 x_0 = c(3,0),
 parameters = SS_oil$two_factor,
 t = 1,
 dt = 1/12,
 N_simulations = 1e3,
 antithetic = TRUE,
 verbose = TRUE)


}
\keyword{internal}
