% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Stitch.Contracts}
\alias{Stitch.Contracts}
\title{Stitch Futures Contracts}
\usage{
Stitch.Contracts(
  Futures,
  TTM = NULL,
  maturity.matrix = NULL,
  rollover.frequency = NULL,
  Contract.Numbers = NULL,
  verbose = FALSE
)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{

Stitch.Contracts(Futures = SS.Oil$Contracts,
TTM = c(1, 5, 9, 13, 17)/12, maturity.matrix = SS.Oil$Contract.Maturities,
rollover.frequency = 1/12, verbose = TRUE)

## ->

stitch_contracts(futures = SS_oil$contracts,
               futures_TTM = c(1, 5, 9, 13, 17)/12,
               maturity_matrix = SS_oil$contract_maturities,
               rollover_frequency = 1/12, verbose = TRUE)

}
\keyword{internal}
