% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{TSFit.Volatility}
\alias{TSFit.Volatility}
\title{Volatility Term Structure of futures returns}
\usage{
TSFit.Volatility(parameters, Futures, TTM, dt)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated due to a change in the name of the function to adhere to the tidyverse style guide.
}
\examples{
V_TSFit <- TSFit.Volatility(
 parameters = SS.Oil$Two.Factor,
 Futures = SS.Oil$Stitched.Futures,
 TTM = SS.Oil$Stitched.TTM,
 dt = SS.Oil$dt)

## ->

V_TSFit <- TSfit_volatility(
 parameters = SS_oil$two_factor,
 futures = SS_oil$stitched_futures,
 futures_TTM = SS_oil$stitched_TTM,
 dt = SS_oil$dt)

}
\keyword{internal}
